+++ inc/template-tags.php
@@ -1557,14 +1557,19 @@ if ( ! function_exists( 'onepress_load_section' ) ) {
          * Hook before section
          */
         do_action('onepress_before_section_' . $section_id);
-        do_action('onepress_before_section_part', $section_id);
+        if ( $section_id != 'hero' ) {
+            do_action('onepress_before_section_part', $section_id);
+        }
+
 
         get_template_part('section-parts/section', $section_id );
 
         /**
          * Hook after section
          */
-        do_action('onepress_after_section_part', $section_id);
+        if ( $section_id != 'hero' ) {
+            do_action('onepress_after_section_part', $section_id);
+        }
         do_action('onepress_after_section_' . $section_id);
     }
 }